Integrate MobX (#44)
* chore(.flowconfig): make Flow work better with MobX

* Prepare for MobX integration

* refactor(router): rewrite to make it work better with MobX

* refactor(<App />): inject root store into the app

Use Mobx’s `<Provider />` component to pass in the root store to the `<App />` component.

* feat(<GlobalHeader />): add GlobalHeader component

* fix(.flowconfig): remove duplication item

* chore: better Flow definition support

* Add common UI components

* Add new deps

* refactor: integrate MobX with the `<Welcome />` screen

* feat(utils): add utils

* refactor(router): rename 'onboarding' → 'wizard'

We’re not really providing an “on-boarding” process. It’s more like a signup wizard. This just feels
right.

BREAKING CHANGE: Routes previously accessed using `/onboarding` now need to be accessed using
`/wizard`.

* feat: add “mock” REST client

“mock” because it really is a promise that resolves itself to true 😂

* feat: add multi-step `<WizardForm />`

BREAKING CHANGE: Previously, the wizard form or the “on-boarding” form was just a large form. Now it
is wrapped using `antd`’s `Form.create()` decorator, allowing us to leverage their in-built support
for form validation.

* refactor: miscellaneous

* Use Mobx React DevTools during development

* fix(<Welcome />): use the current blockchain name instead

* feat: add `BlockchainStore`

This store is used to manage the state in the wizard (form)

closes #25

* Push

* refactor(routes/RouteStore): use the newly created `BlockchainStore`

* Update settings

* refactor: compute dashboard access at component-level instead of root level